Celestial Bodies, Stations, and Factions by System
Added support for querying bodies, stations, and factions for a given
system.
Added test for new methods.
Corrected package name for proguard dto rules.

+++ edsm/src/main/java/com/phapps/elitedangerous/edsm/dto/StationDto.java
@@ -11,7 +11,7 @@ public class StationDto {
     @SerializedName("type")
     private String mType;
     @SerializedName("distanceToArrival")
-    private Integer mDistanceToArrival;
+    private Double mDistanceToArrival;
     @SerializedName("allegiance")
     private String mAllegiance;
     @SerializedName("government")
@@ -49,11 +49,11 @@ public class StationDto {
         mType = type;
     }
 
-    public Integer getDistanceToArrival() {
+    public Double getDistanceToArrival() {
         return mDistanceToArrival;
     }
 
-    public void setDistanceToArrival(Integer distanceToArrival) {
+    public void setDistanceToArrival(Double distanceToArrival) {
         mDistanceToArrival = distanceToArrival;
     }
